Fix styling of scale selector options
