+++ src/h5web/toolbar/controls/ColorMapSelector/ColorMapOption.tsx
@@ -1,0 +1,28 @@
+import {
+  ColorMap,
+  getLinearGradient,
+  INTERPOLATORS,
+} from '../../../../packages/lib';
+import styles from './ColorMapSelector.module.css';
+
+interface Props {
+  option: ColorMap;
+}
+
+function ColorMapOption(props: Props) {
+  const { option } = props;
+  const backgroundImage = getLinearGradient(INTERPOLATORS[option], 'right');
+
+  return (
+    <div className={styles.option}>
+      {option}
+      <div
+        className={styles.gradient}
+        style={{ backgroundImage }}
+        data-keep-colors
+      />
+    </div>
+  );
+}
+
+export default ColorMapOption;
