+++ src/h5web/toolbar/controls/ColorMapSelector/ColorMapSelector.module.css
@@ -2,6 +2,12 @@
   display: flex;
 }
 
+.option {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
 .gradient {
   position: relative;
   top: 1px;
