+++ src/h5web/toolbar/controls/ColorMapSelector/ColorMapSelector.tsx
@@ -1,27 +1,10 @@
-import { INTERPOLATORS } from '../../../vis-packs/core/heatmap/interpolators';
 import type { ColorMap } from '../../../vis-packs/core/heatmap/models';
-import { getLinearGradient } from '../../../vis-packs/core/heatmap/utils';
 import Selector from '../Selector/Selector';
 import styles from './ColorMapSelector.module.css';
 import ToggleBtn from '../ToggleBtn';
 import { FiShuffle } from 'react-icons/fi';
 import { COLORMAP_GROUPS } from './groups';
-
-function ColorMapOption(props: { option: ColorMap }) {
-  const { option } = props;
-  const backgroundImage = getLinearGradient(INTERPOLATORS[option], 'right');
-
-  return (
-    <>
-      {option}
-      <div
-        className={styles.gradient}
-        style={{ backgroundImage }}
-        data-keep-colors
-      />
-    </>
-  );
-}
+import ColorMapOption from './ColorMapOption';
 
 interface Props {
   value: ColorMap;
