+++ src/h5web/toolbar/controls/ScaleSelector/ScaleSelector.module.css
@@ -1,0 +1,8 @@
+.option {
+  display: flex;
+  align-items: center;
+}
+
+.icon {
+  composes: icon from '../../Toolbar.module.css';
+}
