+++ src/h5web/visualizer/Visualizer.module.css
@@ -14,7 +14,7 @@
 
 .visBar {
   position: relative;
-  z-index: 1;
+  z-index: 1; /* for toolbar menus (overflow, selectors) */
   display: flex;
   justify-content: space-between;
   height: 2.5rem;
